requirements-dev.txt
# Include production requirements
-r requirements.txt

# Development Tools
ipython==8.17.2
ipdb==0.13.13
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Code Quality & Linting
black==23.11.0
flake8==6.1.0
isort==5.12.0
pre-commit==3.5.0
bandit==1.7.5

# Testing & Coverage
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
coverage==7.3.2

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Performance Testing
django-silk==5.0.4

# Security Testing
bandit==1.7.5
safety==2.3.5

# Local Development
python-dotenv==1.0.0
django-extensions==3.2.3
